+++ .gitattributes
@@ -10,7 +10,7 @@ go.sum -diff -merge linguist-generated=true
 pkg/security/probe/constantfetch/btfhub/constants.json -diff -merge linguist-generated=true
 pkg/security/seclwin/** -diff -merge linguist-generated=true
 # Fixtures should have LF line endings because they are checked against OCI packages built on Linux
-pkg/fleet/internal/fixtures/** text=auto eol=lf
+pkg/fleet/installer/fixtures/** text=auto eol=lf
 
 # Fix `git diff` when running on the below file formats.
 # Our windows build image uses MinGit which tries to use the astextplain diff algorithm (https://git-scm.com/docs/gitattributes#_setting_the_internal_diff_algorithm).
@@ -27,26 +27,25 @@ pkg/fleet/internal/fixtures/** text=auto eol=lf
 #	textconv = astextplain
 # ```
 
-
-*.doc	diff
-*.DOC	diff
-*.docx	diff
-*.DOCX	diff
-*.docm	diff
-*.DOCM	diff
-*.dot	diff
-*.DOT	diff
-*.dotx	diff
-*.DOTX	diff
-*.dotm	diff
-*.DOTM	diff
-*.pdf	diff
-*.PDF	diff
-*.rtf	diff
-*.RTF	diff
-*.ods	diff
-*.ODS	diff
-*.odf	diff
-*.ODF	diff
-*.odt	diff
-*.ODT	diff
+*.doc diff
+*.DOC diff
+*.docx diff
+*.DOCX diff
+*.docm diff
+*.DOCM diff
+*.dot diff
+*.DOT diff
+*.dotx diff
+*.DOTX diff
+*.dotm diff
+*.DOTM diff
+*.pdf diff
+*.PDF diff
+*.rtf diff
+*.RTF diff
+*.ods diff
+*.ODS diff
+*.odf diff
+*.ODF diff
+*.odt diff
+*.ODT diff
