+++ pkg/fleet/installer/fixtures/fs.go
