+++ pkg/fleet/installer/fixtures/oci-layout-simple-v1-linux2-amd128.tar
