+++ pkg/fleet/installer/fixtures/oci-layout-simple-v1.tar
