+++ pkg/fleet/installer/fixtures/oci-layout-simple-v2.tar
