+++ pkg/fleet/installer/fixtures/server.go
