+++ pkg/fleet/installer/fixtures/simple-v1-config/datadog.yaml.example
