+++ pkg/fleet/installer/fixtures/simple-v1/executable.sh
