+++ pkg/fleet/installer/fixtures/simple-v1/file.txt
