+++ pkg/fleet/installer/fixtures/simple-v2-config/datadog.yaml.example
