+++ pkg/fleet/installer/fixtures/simple-v2/executable-new.sh
